+++ deepctr/estimator/utils.py
@@ -29,29 +29,29 @@ class Head(_Head):
                          unweighted_loss,
                          weights=None):
 
-        labels = tf.to_float(labels)
-        predictions = tf.to_float(predictions)
+        labels = to_float(labels)
+        predictions = to_float(predictions)
 
-        with tf.name_scope(None, 'metrics', (labels, logits, predictions,
+        with name_scope(None, 'metrics', (labels, logits, predictions,
                                              unweighted_loss, weights)):
             metric_ops = {
-                _summary_key(self._name, "prediction/mean"): tf.metrics.mean(predictions, weights=weights,
+                _summary_key(self._name, "prediction/mean"): get_metrics().mean(predictions, weights=weights,
                                                                              name="prediction/mean"),
-                _summary_key(self._name, "label/mean"): tf.metrics.mean(labels, weights=weights, name="label/mean"),
+                _summary_key(self._name, "label/mean"): get_metrics().mean(labels, weights=weights, name="label/mean"),
             }
             if self._task == "binary":
-                metric_ops[_summary_key(self._name, "binary_crossentropy")] = tf.metrics.mean(unweighted_loss,
+                metric_ops[_summary_key(self._name, "binary_crossentropy")] = get_metrics().mean(unweighted_loss,
                                                                                               weights=weights,
                                                                                               name="binary_crossentropy")
-                metric_ops[_summary_key(self._name, "AUC")] = tf.metrics.auc(labels, predictions, weights=weights,
+                metric_ops[_summary_key(self._name, "AUC")] = get_metrics().auc(labels, predictions, weights=weights,
                                                                              name="AUC")
             else:
-                metric_ops[_summary_key(self._name, "mse")] = tf.metrics.mean(unweighted_loss, weights=weights,
+                metric_ops[_summary_key(self._name, "mse")] = get_metrics().mean(unweighted_loss, weights=weights,
                                                                               name="mse")
 
-                metric_ops[_summary_key(self._name, "MSE")] = tf.metrics.mean_squared_error(labels, predictions,
+                metric_ops[_summary_key(self._name, "MSE")] = get_metrics().mean_squared_error(labels, predictions,
                                                                                             weights=weights, name="MSE")
-                metric_ops[_summary_key(self._name, "MAE")] = tf.metrics.mean_absolute_error(labels, predictions,
+                metric_ops[_summary_key(self._name, "MAE")] = get_metrics().mean_absolute_error(labels, predictions,
                                                                                              weights=weights,
                                                                                              name="MAE")
             return metric_ops
@@ -70,7 +70,7 @@ class Head(_Head):
 
     def create_estimator_spec(
             self, features, mode, logits, labels=None, train_op_fn=None):
-        with tf.name_scope('head'):
+        with name_scope('head'):
             logits = tf.reshape(logits, [-1, 1])
             if self._task == 'binary':
                 pred = tf.sigmoid(logits)
@@ -88,7 +88,7 @@ class Head(_Head):
             labels = tf.reshape(labels, [-1, 1])
 
             loss = self.create_loss(features, mode, logits, labels)
-            reg_loss = tf.losses.get_regularization_loss()
+            reg_loss = get_losses().get_regularization_loss()
 
             training_loss = loss + reg_loss
 
@@ -128,9 +128,9 @@ def deepctr_model_fn(features, mode, logits, labels, task, linear_optimizer, dnn
 def get_train_op_fn(linear_optimizer, dnn_optimizer):
     def _train_op_fn(loss):
         train_ops = []
-        global_step = tf.train.get_global_step()
-        linear_var_list = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES, LINEAR_SCOPE_NAME)
-        dnn_var_list = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES, DNN_SCOPE_NAME)
+        global_step = tf.train.get_global_step() if tf.__version__ < "2.0.0" else tf.compat.v1.train.get_global_step()
+        linear_var_list = get_collection(get_GraphKeys().TRAINABLE_VARIABLES, LINEAR_SCOPE_NAME)
+        dnn_var_list = get_collection(get_GraphKeys().TRAINABLE_VARIABLES, DNN_SCOPE_NAME)
         if len(dnn_var_list) > 0:
             train_ops.append(
                 dnn_optimizer.minimize(
@@ -144,13 +144,56 @@ def get_train_op_fn(linear_optimizer, dnn_optimizer):
 
         train_op = tf.group(*train_ops)
         with tf.control_dependencies([train_op]):
-            return tf.assign_add(global_step, 1).op
+            return tf.assign_add(global_step, 1).op if tf.__version__ < "2.0.0" else tf.compat.v1.assign_add(global_step, 1).op
 
     return _train_op_fn
 
 def variable_scope(name_or_scope):
-    if tf.__version__ >= "2.0.0":
-        return tf.compat.v1.variable_scope(name_or_scope)
-    else:
+    if tf.__version__ < "2.0.0":
         return tf.variable_scope(name_or_scope)
+    else:
+        return tf.compat.v1.variable_scope(name_or_scope)
 
+def name_scope(name, default_name=None, values=None):
+    if tf.__version__ < "2.0.0":
+        return tf.name_scope(name,default_name,values)
+    else:
+        return tf.compat.v1.name_scope(name,default_name,values)
+
+def get_collection(key, scope=None):
+    if tf.__version__ < "2.0.0":
+        return tf.get_collection(key,scope=scope)
+    else:
+        return tf.compat.v1.get_collection(key,scope=scope)
+
+def get_GraphKeys():
+    if tf.__version__ < "2.0.0":
+        return tf.GraphKeys
+    else:
+        return tf.compat.v1.GraphKeys
+
+def get_losses():
+    if tf.__version__ < "2.0.0":
+        return tf.losses
+    else:
+        return tf.compat.v1.losses
+
+
+def input_layer(features,feature_columns):
+    if tf.__version__ < "2.0.0":
+        return tf.feature_column.input_layer(features, feature_columns)
+    else:
+        return tf.compat.v1.feature_column.input_layer(features, feature_columns)
+
+def get_metrics():
+    if tf.__version__ < "2.0.0":
+        return tf.metrics
+    else:
+        return tf.compat.v1.metrics
+
+
+def to_float(x,name="ToFloat"):
+    if tf.__version__ < "2.0.0":
+        return tf.to_float(x,name)
+    else:
+        return tf.compat.v1.to_float(x,name)